AgentFinder: extract match rendering into AgentMatchWindow.tsx
AgentFinder.tsx was 1025 lines serving too many responsibilities.
Extracted everything related to rendering matched message content:
- BubbleLayout + getAssistantBubbleProps/getUserBubbleProps/getAgentSentBubbleProps
- Snippet windowing logic (findMatchIndexes, createSnippetWindow, etc.)
- HighlightedSnippet, MatchWindow, MatchMessage, MatchPair components

AgentFinder.tsx: 1025 → 707 lines, focused on finder state and keyboard behavior.
AgentMatchWindow.tsx: new file, ~340 lines, focused on match display.
